Allow quotes from multiple messages in a single reply
When replying to a message, we use the existing inline composer if it
already references the message we're replying to, or if we're replying
with a quote.  Replies without quotes get made into new composers.

When adding a new message as a reply in a composer, we add the relevant
addresses to the to and cc lines, based on the current reply type.  We
remove duplicates, and don't include any to addresses as cc.  We also
keep track of a total to and total cc list, which gets swapped in if a
reply action is to change the reply type.

We keep track of all the replied-to message ids and include all of them
in the In-Reply-To header.  However, the References header only includes
message ids from the first message replied to.  This is because RFC 2822
requires this header to represent a single lineage, which we cannot
guarantee if we're replying to different branches of the conversation.