Improve folder full-normalization time using IMAP SEARCH
When messages have been appended and removed from a folder, Geary has
no choice but to do a full UID scan of the server to figure out
exactly what's changed.  Prior code use a FETCH command, but SEARCH is
a better choice because there's less overhead in a SEARCH result: only
the UIDs are returned and they may be returned in a single line or in
batches.  Also, there's no reason to return those results in a sorted
set, which was driving up transaction time, so an unsorted set is
returned.

On my account, a full normalization of All Mail went from 8s to 1s
with this change.