Fix app & desktop becoming unresponsive when clicking the gear menu.
When the shell is not showing the app menu, we are adding a gear menu and
using that for the app menu to work around an issue in Mint Cinnamon
always showing a menu bar when an app menu was present. However under
gnome-shell, this causes both Geary and gnome-shell to both start fully
consuming the CPU when the gear menu was activated.

Perhaps GtkApplication{Window} was notifying or querying shell about the
state of the actions since they are in the in the "app" namespace, but
shell not knowing or caring about them was returning an error, causing
Geary to retry?

Removing the gear menu and letting GTK+ add its own app menu to the
header when shell is not showing it fixes the issue.

We could load the app_menu actions into the "win" namespace, but that
would require adding yet more workarounds (at least a duplicate
app_menu.ui file or something else similary sucky) for someone else's bug
and for users that will never see or need the fix.

Bug 759980.

* src/client/application/geary-controller.vala: Always set the
  application's app_menu.

* src/client/components/main-toolbar.vala,
  src/client/composer/composer-headerbar.vala: Remove gear menu.

* src/client/components/main-window.vala: Set show-menubar to false
  so we either get the shell managed app-menu or GTK managed one
  in the header bar.