Fix missing text displaying an image with text parts and inline images.
When Geary.RFC822.Message::construct_body_from_mime_parts is looking to
build a message body to display, during the first pass that looks for
HTML content the plain text part is ignored but the inline image gets
converted to a HTML IMG element and returned as the body. This makes it
look like an HTML part was found, and hence the second pass looking for
plain text never takes place.

Fix this by not taking a wild stab at the body type - check to see if it
has HTML parts and if so just het them, if not then get plain text parts.

Bug 767438.

* src/engine/rfc822/rfc822-message.vala (Message): Add ::has_html_body
  and ::has_plain_body to allow callers to query available body content
  types before attempting to get them. Remove ::get_body since it's
  making a policy decision that callers should be making instead, fix
  call sites.