Rework ComposerWidget's pending attachment implementation. Bug 712995.
This merges the pending attachment add & check methods to do both at
once, so the logic for determining if a panding attachment should be
added doesn't need to be duplicated.

Also changes the behaviour of pending attachments to always include
inline attachments, even when only replying, since they should be
displayed if included in the quote.

* src/client/composer/composer-widget.vala (ComposerWidget): Add state
  flag to specify which pending attachments should be included. Move all
  private attachment-related fields to the same place in the
  source. Combine ::add_pending_attachments and
  ::check_pending_attachments into new ::update_pending_attachments
  method, update call sites.