Use preview synthesis when conversation list is refreshing previews.
Since currently, PREVIEW has some bad assumptions, we want to avoid using
it. As a workaround, when listing messages with PREVIEW, include ALL so
that HEADER and BODY are also retreived and the preview can synthesised,
also NONE to ensure that after setting ALL, UIDs will be returned in the
result.

Bug 714317.

* src/client/conversation-list/conversation-list-store.vala
  (ConversationListStore): Include ALL and NONE in WITH_PREVIEW_FIELDS.

* src/client/conversation-viewer/conversation-list-box.vala
  (ConversationListBox): Remove PREVIEW from REQUIRED_FIELDS, it's no
  longer needed.