Workaround xgettext not recognising Elementary Contractor files.
We can't just simply run xgettext twice (once as normal, once with just
the contractor file and passing "--desktop" as an arg, since
l10n.gnome.org won't pick up the second pass and the contract will remain
untranslated.

So work around by renaming the contractor file such that it is
recognised, then rename it back to normal when translating it.

* desktop/geary-attach.contract.in: Renamed to
  desktop/geary-attach.contract.desktop.in.

* desktop/CMakeLists.txt: Undo the rename when translating.

* po/CMakeLists.txt, po/Makevars: Add a xgettext keyword for the
  contractor Description field.

* po/POTFILES.in: Chase the file rename.