Don't add IMAP connections back to the pool when the account is closing.
This updates the client session manager to close IMAP connections without
waiting when they are released and the account is closing, rather than
adding them back to the pool again. Sessions that are in the Selected
state due to being used by a folder are also simply closed, rather than
attempting to return them to Authenticated (i.e. not selected) state
first.

This means when a folder is closed and releases its session that it does
not block if its session has hung, making application shutdown less
likely to hang as well, helping Bug 745561 along a bit further.