Explicitly close the Serializer's buffer stream.
Since the Serializer is passed a buffered GLib stream that uses a
background thread to manage async IO, it's better to explicitly close the
stream so we know the thread will be shut down at that point, rather than
at some random point in the future.

This will hopefully fix Bug 792219, but it's hard to know since it is a
rarely seen race.

* src/engine/imap/transport/imap-client-connection.vala (Connection):
  Keep a reference to the BufferedOutputStream we pass the serialiser,
  and explicitly close and dispose after disposing of the serialiser
  instance.