Remove GenericAccount's IMAP account session attribute.
For most of the time, GenericAccount does not need an open client
session, yet it always claims one from the pool on open and hangs on to
it. This means that the account synchroniser for example cannot use the
same session to perform sync, requiring an additional session to be
opened.

This patch removes the held session with methods to claim and release
account sessions instances for those code paths that need it, allowing
the previously claimed session to be re-used when not in use.

* src/engine/imap-engine/imap-engine-generic-account.vala
  (GenericAccount): Remove remote_open_lock and remote_session, just
  claim new client sessions from the pool as needed instead. Add
  release_account_session to allow claimed sessions to be returned to the
  pool. Update uses of remote_session to claim and release instead.

* src/engine/imap/transport/imap-client-session-manager.vala
  (ClientSessionManager): Update ready signal to include a parameter
  indicating if ready or not, so GenericAccount can still block
  claim_account_session callers when not ready. Rework all code that sets
  is_ready to use a common notify method that both does that and fires
  the signal, and rework all code paths that stop the pool to use a
  common code path that uniformly resets timers, notifies no longer
  ready, and drops all client sessions.

* src/engine/imap/api/imap-account-session.vala: Remove distinction
  between fetch_folder_async and fetch_folder_cached_async
  since account session objects are now short-lived.